# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "round-1",
        "description": "round-1 doc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/testimonial": {
            "get": {
                "tags": [
                    "testimonial"
                ],
                "summary": "Get all testimonials",
                "description": "Endpoint to Get all testimonials",
                "operationId": "5df40b6818d8707213a2bcd12693116f",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "testimonial"
                ],
                "summary": "Create a testimonial for a product",
                "description": "Endpoint to create a testimonial with product, text, image, and video",
                "operationId": "2441feb066aef91d61e8ef95f16277c3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "text"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID of the product for the testimonial",
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "text": {
                                        "description": "Testimonial text",
                                        "type": "string",
                                        "example": "This product changed my life!"
                                    },
                                    "image": {
                                        "description": "Testimonial image file (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "video": {
                                        "description": "Testimonial video file (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/testimonial/{id}": {
            "get": {
                "tags": [
                    "testimonial"
                ],
                "summary": "Get a testimonial by id",
                "description": "Endpoint to Get a testimonial by id",
                "operationId": "2534411137189ebfb36744abc5191a95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "testimonial"
                ],
                "summary": "Update a testimonial for a product",
                "description": "Endpoint to update a testimonial with product, text, image, and video",
                "operationId": "bb7158cf9c4fe35db59d017dc637ac8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "text"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID of the product for the testimonial",
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "text": {
                                        "description": "Updated testimonial text",
                                        "type": "string",
                                        "example": "This product is even better now!"
                                    },
                                    "image": {
                                        "description": "Updated testimonial image file (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "video": {
                                        "description": "Updated testimonial video file (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "testimonial"
                ],
                "summary": "Delete testimonial by id",
                "description": "Endpoint to Delete testimonial by id",
                "operationId": "850c769b6c7bfe810c4158ed4982d797",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "required": [
                    "status",
                    "message"
                ],
                "properties": {
                    "status": {
                        "description": "Status of the API response",
                        "type": "string"
                    },
                    "message": {
                        "description": "Message describing the result",
                        "type": "string"
                    },
                    "data": {
                        "description": "Optional data payload",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}